+++ src/js/defaultOptions.js
@@ -5,7 +5,7 @@ export const defaultOptions = {
   buttonCallback: null,
   buttonText: null,
   duration: 4000,
-  message: "Ackbar snackbar: It's a snackbar",
+  message: "Ackbar snackbar: It's a snackbar!",
   size: 'normal',
   type: 'auto',
   variant: 'default',