+++ src/styles/ackbar-bar.styles.js
@@ -36,44 +36,39 @@ export const ackbarBarStyles = css`
   :host([size='small']) {
     padding: calc(var(--ackbar-padding-normal) / 3)
       calc(var(--ackbar-padding-normal) / 2);
-    font-size: 87.5%;
-  }
-
-  :host([size='small']) .ackbar-bar__button {
-    font-size: 87.5%;
   }
 
   :host([size='small']) .ackbar-bar__message,
   :host([size='small']) .ackbar-bar__button {
-    font-size: 0.75rem;
+    font-size: calc(var(--ackbar-base-font-size) * 0.75);
   }
 
   :host([size='large']) {
     padding: calc(var(--ackbar-padding-normal) * 0.5)
       calc(var(--ackbar-padding-normal) * 1.25);
-    font-size: 125%;
   }
 
+  :host([size='large']) .ackbar-bar__message,
   :host([size='large']) .ackbar-bar__button {
-    font-size: 125%;
+    font-size: calc(var(--ackbar-base-font-size) * 1.5);
   }
 
   /* HAS BUTTON */
   :host([hasButton]) {
     display: grid;
     grid-template-columns: 1fr auto;
-    grid-gap: 1rem;
+    grid-gap: var(--ackbar-grid-gap);
   }
 
   /* BUTTON */
   .ackbar-bar__button {
     border: none;
-    background-color: transparent;
-    font-size: 1rem;
+    background-color: var(--ackbar-color-button-background);
+    font-size: var(--ackbar-base-font-size);
     color: var(--ackbar-color-button-default);
     outline: none;
-    opacity: 0.75;
-    padding: 0.125rem 0.5rem;
+    opacity: var(--ackbar-button-opacity);
+    padding: var(--ackbar-button-padding);
     border-radius: 0.25rem;
   }
 
@@ -86,14 +81,14 @@ export const ackbarBarStyles = css`
 
   .ackbar-bar__message {
     margin: 0;
-    font-size: 1rem;
+    font-size: var(--ackbar-base-font-size);
   }
 
   @media (hover: hover) {
     .ackbar-bar__button:hover {
-      background-color: var(--ackbar-color-button-hover-background);
+      background-color: var(--ackbar-color-button-background-hover);
       cursor: pointer;
-      opacity: 1;
+      opacity: var(--ackbar-button-opacity-hover);
     }
   }
 `;