+++ changelog.md
@@ -7,3 +7,11 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Improve styles to make the snackbars responsive
+
+## [0.0.6] - 2021-02-22
+### Added
+- css variable for the base font size
+- css variable for the button padding
+- css variable for the button background color
+- css variable for button opacity (default and on hover)
+- add variable for the grid gap